functions/submit_job/requirements.txt
google-cloud-firestore>=2.0.0
flask>=3.0.0
uuid>=1.30 